[MacOS] Share LabelRenderer with iOS
